KYAA-16 Restructure and make tests pass for now
