+++ app/models/abstractAnimeDataSource.js
@@ -1,0 +1,9 @@
+module.exports = class AbstractAnimeDataSource {
+    getAnime(name) {
+        throw "Implementation error";
+    }
+
+    _constructAnimeModel(animeResponse) {
+        throw "Implementation error";
+    }
+}
