+++ app/models/slackEvents.js
@@ -1,5 +1,5 @@
 const slackEventsApi = require('@slack/events-api');
-const { MyAnimeList } = require('./data');
+const MyAnimeList = require('./myAnimeListClient');
 const Entities = require('html-entities').AllHtmlEntities;
 const popura = require('popura');
 const malClient = new MyAnimeList(popura(process.env.MAL_USERNAME, process.env.MAL_PASSWORD), new Entities());
