+++ index.js
@@ -8,8 +8,8 @@ const bodyParser = require('body-parser');
 const slackEventsApi = require('@slack/events-api');
 const { WebClient } = require('@slack/client');
 
-const SlackClient = require('./slackClient');
-const { registerEvents } = require('./slackEvents');
+const SlackClient = require('./app/models/slackClient');
+const { registerEvents } = require('./app/models/slackEvents');
 
 const slackEventsAdapter = slackEventsApi.createSlackEventAdapter(process.env.SLACK_VERIFICATION_TOKEN, { includeBody: true });
 const slackWebClient = new WebClient(process.env.SLACK_TEAM_OAUTH_TOKEN);
